Move the Python files containing configuration into the new configuration directory
