+++ src/lib/configuration_objects.py
@@ -1,0 +1,34 @@
+class CodingModes(object):
+    SINGLE = "SINGLE"
+    MULTIPLE = "MULTIPLE"
+
+
+class CodingConfiguration(object):
+    def __init__(self, coding_mode, code_scheme, coded_field, fold_strategy, analysis_file_key=None, cleaner=None):
+        assert coding_mode in {CodingModes.SINGLE, CodingModes.MULTIPLE}
+
+        self.coding_mode = coding_mode
+        self.code_scheme = code_scheme
+        self.coded_field = coded_field
+        self.analysis_file_key = analysis_file_key
+        self.fold_strategy = fold_strategy
+        self.cleaner = cleaner
+
+
+# TODO: Rename CodingPlan to something like DatasetConfiguration?
+class CodingPlan(object):
+    def __init__(self, raw_field, coding_configurations, raw_field_fold_strategy, coda_filename=None, ws_code=None,
+                 time_field=None, run_id_field=None, icr_filename=None, id_field=None, code_imputation_function=None):
+        self.raw_field = raw_field
+        self.time_field = time_field
+        self.run_id_field = run_id_field
+        self.coda_filename = coda_filename
+        self.icr_filename = icr_filename
+        self.coding_configurations = coding_configurations
+        self.code_imputation_function = code_imputation_function
+        self.ws_code = ws_code
+        self.raw_field_fold_strategy = raw_field_fold_strategy
+
+        if id_field is None:
+            id_field = "{}_id".format(self.raw_field)
+        self.id_field = id_field
